Add: multi-term and its setup, setup-work
Added multi-term package and its associated setup.

Removed secrets.el.gpg and replaced it with setup-work.
The require for 'setup-work does not result in error if that package is
not found. The 'setup-work' is meant to be a local file containing work
related stuff and not for sharing on github.