Major: Refactoring to use `use-package`
It was decided to use `use-package` instead of `req-package`. Faced a
lot of issues when using `req-package`. If I did `req-package for a
package like `hydra` and if `defhydra` were used in other "req-package"d
setups, it wouldn't work. I would need to use `use-package` or `require`
for `hydra`.

For consistency and simplicity, as I didn't see much advantage using
`req-package`, I am moving to using just `use-package`.