Add natural language to org html export postambles
The prepositions "by" and "on" and extra spaces are now added
automatically based on the org document's #+AUTHOR and #+DATE being
nil or non-nil.

For example, if both date and author are non-nil, the postamble will
look like:

"Exported using Emacs 25.0.50.1 (Org mode 8.3.3) by Kaushal Modi on Jan
27 2016, Wed."

But if author were nil and date were non-nil, it would look like:

"Exported using Emacs 25.0.50.1 (Org mode 8.3.3) on Jan 27 2016, Wed."

And similarly for the remaining two scenarios.