Make the region-or-whole functions robust
Earlier when using the advice to get the "region or whole"
functionality worked for `indent-region` but ONLY if (mark) was
non-nil in the buffer. It failed if (mark) returned nil for a buffer.

The reason is that the `interactive` function in `indent-region` has an
`r` argument. In `call-interactively` C function, if the argument is
`r`, error is signaled if mark is nil or undefined for that buffer. This
error signaling is done even before nadvice can set
the (point-min)/(point-max) region arguments using :around or
:filter-args.

The new approach simply defines wrapper functions that pass
the (point-min)/(point-max) args if mark is not active (or region is not selected).