Use (use-region-p) instead of mark-active
As per the manual node "(elisp) The Mark", this variable should not be
used directly.

The mark is active when this variable is non-‘nil’.  This variable
is always buffer-local in each buffer.  Do _not_ use the value of
this variable to decide whether a command that normally operates on
text near point should operate on the region instead.  Use the
function ‘use-region-p’ for that (*note The Region::).