Re-wrote extend-rectangle-to-end
The extend-rectangle-to-end function from the rectangle-utils did not
work correctly. There was a bug in the goto-longest-region-line function
in that package.

So instead, I took few elements from the original
extend-rectangle-to-end function and re-wrote it to get it working
finally!

As that was the only function I needed from the rectangle-utils package,
I have uninstalled that package.

The modi/extend-rectangle-to-end also supports regions where the point
and mark are on the same column, including the case where the end of the
region is on the beginning of a line (0th column). So you are allowed to
"C-a C-n |" to quickly create a 2-row rectangle. (Here "|" is a binding
to modi/extend-rectangle-to-end in region-bindings-mode-map).