Bind <f1> to goto-line; remap it to modi/goto-line
This is a cleaner approach so that I do not need to hard-code <f1> to
something like modi/goto-line. The reason is that then I can remap
goto-line to pdf-view-goto-page in pdf-view-mode instead of having to
remap modi/goto-line.

Now if avy is not loaded, f1 will call goto-line in non-pdf-view major
modes, while call pdf-view-goto-page in pdf-view-mode.

While, if avy is loaded, the same binding will call modi/goto-line in
non-pdf-view major modes, but STILL call pdf-view-goto-page in pdf-view-mode.