Set org-edit-src-content-indentation to 0
I did not like the default 2 space indentation added on exiting the *Org
Src* buffer. That created aesthetic inconsistency because if I did not
use «C-c '» to edit a BEGIN_SRC block, I would not no indentation. But
if I did edit the source block using that, I would get the 2 space
indentation.

So for consistency, I am setting that indentation to 0.