Now scroll other window works even if that has pdf
Earlier the scroll-other-window did not work if the other window had a
PDF file (pdf-view-mode) and the window displayed the full page.

This advice now calls the
pdf-view-next-line-or-next-page/pdf-view-previous-line-or-previous-page
functions in that PDF window so that if that window is showing the full
page, the next/previous page is shown. If that window is NOT showing the
full page, linewise scroll will happen.