Always require setup-counsel
The counsel commands are useful even in the cases where user has chosen
to use ido instead of ivy by default (Example: counsel-ag). So always
require it regardless of the value of the disable-pkg-ivy var.

In addition, bind C-x C-f, M-x, etc bindings to counsel commands only if
disable-pkg-ivy var is not set to a non-nil value.