Bind windmove commands to s- prefix instead of S-
With shift-select-mode enabled, a user might need to tweak the text
selection using Shift + arrow keys (Note that that would work only if
the hardcore-mode is disabled, because having that mode enabled will
prevent the use of arrow keys).  But that was not possible as the
S-<arrow> keys were bound to the windmove-* commands.

Those have now been changed to the s-<arrow> keys (Windows key + arrow
keys). Now a user can have both shift-select-mode and windmove bindings
functionalities (when hardcode-mode is disabled).