Do not show "./" and "../" during C-x C-f
I had already set `counsel-find-file-ignore-regexp` to not show "./" and
"../" during `counsel-find-file`. But at some point that functionality
changed in ivy and now we also need to set `ivy-extra-directories` so
that it does not contain "./" and "../". So now we also need to set
`ivy-extra-directories` to a value like nil.

C-x C-f looks clean once again :)