Further binding cleanup "C-x m p" → "C-x m P"
I have rarely needed to print buffers to PDF files. But I almost never
print them to actual printers on paper. So there's no need to waste a
key binding for that. So I am removing the binding for
`ps-print-buffer-with-faces`.

Now `modi/pdf-print-buffer-with-faces` is bound to "C-x m P",
and "C-x m p" is freed up.