Move clipboard tweaks to setup-editing.el
These setting were moved out of setup-misc.el to setup-editing.el so
that a user can override these settings in setup-personal.el.

Order of loading these files:

  ...
  setup-editing
  ...
  setup-personal (this is where a user can override anything set before
                  this point)
  ...
  setup-misc (this is required at the very end)

Also the value of trash-directory is changed so that it works across
systems (Windows/Linux). Now we make use of the
`temporary-file-directory` variable instead of hard-coding "/tmp/trash".