Cleaner way of quitting emacs without desktop save
It is necessary to disable desktop-save-mode if user chooses to quit
emacs without saving the desktop.

Otherwise the desktop-kill function in the emacs-kill-hook will ask the
user if they want to save the desktop if the session is modified and
needs to be saved.

bound-and-true-p is not needed as desktop-save-mode autoloads
desktop.el.