Minor tweak to endless/narrow-or-widen-dwim
Do NOT delete other windows after calling org-edit-src-code; I like to
view the org document and the src edit buffer in split windows.

Also "t" in "(ignore-errors (org-edit-src-code) t)" is added as a
fail-safe means so that in the event of no errors, that form returns t and
the cond case is entered.