Don't recalc org tables during export
modi/org-table-recalculate-buffer-tables is already added to
before-save-hook. I always save the file before exporting. So there is
no need adding the same to the org-export-before-processing-hook too.

Another reason for removing this is that I would need a construct like
below in all the org files where I would want to change the buffer local
variable modi/org-table-enable-buffer-wide-recalculation's value which
enables the modi/org-table-recalculate-buffer-tables function.

    #+BIND: foo nil
    # Local Variables:
    # org-export-allow-bind-keywords: t
    # End:

It's not worth that hassle, when anyways I recalculate the tables when
saving .. (Sour grapes? May be :))