Add modi/shell-region
This command is bound to C-x C-e ONLY if the current major mode is
sh-mode and if a region is selected.

Simply select a region in a shell-script file and hit C-x C-e. The
selected code will be run using the shell referenced by that buffer
local variable sh-shell-file. So if the script is a tcsh shell script,
the region will be evaluated using tcsh, and so on.