Bug fix: GUI/terminal mode related logic
The bug was that (display-graphic-p) was evaluated as nil in the init.el
when emacs was launched using "emacsclient -c". When launching frame
using emacsclient, the frame is always created AFTER the init.el is
evaluated. So emacs will always think that there is no frame.

So now the (display-graphic-p) is evaluated in
after-make-frame-functions hook. That way the GUI based code is
correctly activated whether I launch emacs using emacs& or emacsclient -c&