Fix flyspell-correct-ivy not getting autoloaded
The issue was interesting .. I was trying to autoload
flyspell-correct-ivy package by using the :bind keyword. My idea was
that when I hit F12, the flyspell-correct-word-generic command would be
called and that would autoload flyspell-correct-ivy.

But that doesn't actually work like that. It's because
flyspell-correct-word-generic is defined in the flyspell-correct
package, *not* the flyspell-correct-ivy package. As a result, when I hit
F12, only the flyspell-correct package got auto-loaded, not
flyspell-correct-ivy. Looks like flyspell-correct-ivy does not have any
functions that can autoload it. So now I use the :after keyword magic of
use-package. Using that flyspell-correct-ivy gets loaded *only* after
flyspell-correct gets loaded (and flyspell-correct gets loaded when I
hit F12).

:)