Update org-capture-templates with change in API
Starting the commit in
http://orgmode.org/cgit.cgi/org-mode.git/commit/?id=5485170263a46fa17db67b0324c4c4e48bcdfe49
on June 23 2016 on the org master branch, sexp's are not allowed in the
templates. Instead of bare sexp's, we need to provide a function or
variable name if not the absolute/relative (to org-directory) file name
string.

In any case, the files I had in the template were in the
org-directory. So I did not need the (expand-file-name ..) form.

I also removed the personal uvm.org reference from the template.