Fix switch to scratch function
If this function is called from a buffer that is read-only (even if the
major mode is not read-only by default), it will launch an org-mode
scratch buffer. That is not what I intended.

I intend to open org mode scratch buffer ONLY if the major mode is
inherently a read-only mode, like help-mode. This commit fixes that and
gives this intended behavior.