Use rg (ripgrep) for projectile file caching
Use rg (ripgrep) instead of ag if available

https://github.com/BurntSushi/ripgrep

From my use cases, the speeds of rg and ag are about the same.
But rg allows whitelisting files in .ignore files.

So while all the dot files are ignored by default, I can have "!.foo"
to *not* ignore the `.foo` files.