Add modi/verilog-identifier-pcre;fix modi/verilog-find-parent-module
The `modi/verilog-find-parent-module' function runs ag to search for the parent
module. So we need a PCRE expression.

modi/verilog-identifier-pcre is the PCRE version of the Elisp regexp in
modi/verilog-identifier-re.

This commit also fixes the earlier incorrect greedy expression for skipping port
parameter lists: "[^.]" -> "([^.])*?"

Also fix the unintentional global declaration of `module-instance-pcre'; now
that is let-bound.