Make the window spawning less jarring
Earlier if F9 was hit in the code window, eshell opened up in the window where
there was code and code buffer would get buried one level down.

After this fix, the code buffer stayed in the same window, compilation buffer
opens in a side window, then that buffer gets buried and eshell opens up in that
safe window. So the code window stays as it was.

But now as eshell is opened in a different windows than the code window, we need
to ensure that the default-directory for eshell is set to that of the code
window, because if the code foo.nim got compiled to ./foo, eshell needs to open
up in the directory containing ./foo directly.