Use shell-quote-argument instead of identity
As modi/ag-arguments contains "--ignore" "#*#", we need to shell-quote the "#*#"
portion. In general, it is safer to always shell quote strings used to create a
command line command.

But as "%s" is specific to counsel-ag and counsel-rg, do NOT shell-quote that!
For that reason, the " %s" portion is separately concatenated at the very end
instead of passing it to mapconcat.

Also it looks like there is no need to backquote ag and rg.. so removing that.