Add new variable modi/eww-search-alist
modi/eww--go-to-first-search-result now uses this new variable to check if the
eww-search-prefix has a corresponding regexp in there to help find the first
search result.

If modi/eww-search-alist is missing an entry for eww-search-prefix, Google
search is used by default.

DuckDuckGo is not added to modi/eww-search-alist because shr-probe-and-copy-url
fails copying the duckduckgo.com search result URLs when point is on them.

Fixes https://github.com/kaushalmodi/.emacs.d/issues/27